requirements.txt
azure-functions
beautifulsoup4
pandas
requests
sqlalchemy
pyodbc
